requirements.txt
anthropic==0.18.1
requests==2.31.0
flask==3.0.0
python-dotenv==1.0.0
pandas==2.1.4
plotly==5.18.0
